+++ tales_django/entities/Tracks/views/rubric_details_view.py
@@ -2,6 +2,7 @@ from django.http import HttpRequest
 from django.shortcuts import get_object_or_404, render
 
 from core.logging import getDebugLogger
+from tales_django.core.model_helpers import check_required_locale
 from tales_django.core.pages import get_favorites_list_context, get_rubric_tracks_list_context, get_common_context
 from tales_django.entities.Tracks.models import Rubric
 
@@ -10,6 +11,8 @@ logger = getDebugLogger()
 
 
 def rubric_details_view(request: HttpRequest, rubric_id):
+
+    check_required_locale(request)
 
     rubric = get_object_or_404(Rubric, id=rubric_id)