+++ tales_django/entities/Tracks/views/tag_index_view.py
@@ -2,10 +2,14 @@ from django.http import HttpRequest
 from django.shortcuts import render
 from django.utils.translation import gettext_lazy as _
 
+from tales_django.core.model_helpers import check_required_locale
 from tales_django.core.pages import get_favorites_list_context, get_tags_list_context, get_common_context
 
 
 def tag_index_view(request: HttpRequest):
+
+    check_required_locale(request)
+
     context = {
         **get_common_context(request),
         **get_favorites_list_context(request),